+++ CHANGES.rst
@@ -2,6 +2,14 @@
 CHANGES
 =======
 
+----------
+0.2.0-dev0
+----------
+
+* Adds vagrant-based integration test to test compactor against reference libprocess.
+
+* Fixes Python 3 support, pinning to protobof >= 2.6.1 < 2.7 which has correct support.
+
 -----
 0.1.3
 -----